+++ Source/ACE.Entity/AceObject.cs
@@ -203,6 +203,12 @@ namespace ACE.Entity
             set { SetDataIdProperty(PropertyDataId.PhysicsEffectTable, value); }
         }
 
+        public uint? CombatTableDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.CombatTable); }
+            set { SetDataIdProperty(PropertyDataId.CombatTable, value); }
+        }
+
         public uint? PhysicsState
         {
             get { return GetIntProperty(PropertyInt.PhysicsState); }
@@ -465,7 +471,7 @@ namespace ACE.Entity
         public uint? PlacementPosition
         {
             get { return GetIntProperty(PropertyInt.PlacementPosition); }
-            set {SetIntProperty(PropertyInt.PlacementPosition, value); }
+            set { SetIntProperty(PropertyInt.PlacementPosition, value); }
         }
 
         public float? Translucency
@@ -496,6 +502,546 @@ namespace ACE.Entity
         {
             get { return GetDataIdProperty(PropertyDataId.ClothingBase); }
             set { SetDataIdProperty(PropertyDataId.ClothingBase, value); }
+        }
+
+        public uint? CharacterOptions1
+        {
+            get { return GetIntProperty(PropertyInt.CharacterOptions1); }
+            set { SetIntProperty(PropertyInt.CharacterOptions1, value); }
+        }
+
+        public uint? CharacterOptions2
+        {
+            get { return GetIntProperty(PropertyInt.CharacterOptions2); }
+            set { SetIntProperty(PropertyInt.CharacterOptions2, value); }
+        }
+
+        public uint? TotalLogins
+        {
+            get { return GetIntProperty(PropertyInt.TotalLogins); }
+            set { SetIntProperty(PropertyInt.TotalLogins, value); }
+        }
+
+        public uint? AccountIID
+        {
+            get { return GetInstanceIdProperty(PropertyInstanceId.Account); }
+            set { SetInstanceIdProperty(PropertyInstanceId.Account, value); }
+        }
+
+        public bool? IsDeleted
+        {
+            get { return GetBoolProperty(PropertyBool.IsDeleted); }
+            set { SetBoolProperty(PropertyBool.IsDeleted, value); }
+        }
+
+        public ulong? DeleteTime
+        {
+            get { return GetInt64Property(PropertyInt64.DeleteTime); }
+            set { SetInt64Property(PropertyInt64.DeleteTime, value); }
+        }
+
+        public ulong? AvailableExperience
+        {
+            get { return GetInt64Property(PropertyInt64.AvailableExperience); }
+            set { SetInt64Property(PropertyInt64.AvailableExperience, value); }
+        }
+
+        public ulong? TotalExperience
+        {
+            get { return GetInt64Property(PropertyInt64.TotalExperience); }
+            set { SetInt64Property(PropertyInt64.TotalExperience, value); }
+        }
+
+        public uint? Age
+        {
+            get { return GetIntProperty(PropertyInt.Age); }
+            set { SetIntProperty(PropertyInt.Age, value); }
+        }
+
+        public string DateOfBirth
+        {
+            get { return GetStringProperty(PropertyString.DateOfBirth); }
+            set { SetStringProperty(PropertyString.DateOfBirth, value); }
+        }
+
+        public uint? AvailableSkillCredits
+        {
+            get { return GetIntProperty(PropertyInt.AvailableSkillCredits); }
+            set { SetIntProperty(PropertyInt.AvailableSkillCredits, value); }
+        }
+
+        public uint? TotalSkillCredits
+        {
+            get { return GetIntProperty(PropertyInt.TotalSkillCredits); }
+            set { SetIntProperty(PropertyInt.TotalSkillCredits, value); }
+        }
+
+        public uint? NumDeaths
+        {
+            get { return GetIntProperty(PropertyInt.NumDeaths); }
+            set { SetIntProperty(PropertyInt.NumDeaths, value); }
+        }
+
+        public uint? DeathLevel
+        {
+            get { return GetIntProperty(PropertyInt.DeathLevel); }
+            set { SetIntProperty(PropertyInt.DeathLevel, value); }
+        }
+
+        public uint? VitaeCpPool
+        {
+            get { return GetIntProperty(PropertyInt.VitaeCpPool); }
+            set { SetIntProperty(PropertyInt.VitaeCpPool, value); }
+        }
+
+        public bool? IsAdmin
+        {
+            get { return GetBoolProperty(PropertyBool.IsAdmin); }
+            set { SetBoolProperty(PropertyBool.IsAdmin, value); }
+        }
+
+        public bool? IsEnvoy
+        {
+            get { return GetBoolProperty(PropertyBool.IsSentinel); }
+            set { SetBoolProperty(PropertyBool.IsSentinel, value); }
+        }
+
+        public bool? IsArch
+        {
+            get { return GetBoolProperty(PropertyBool.IsArch); }
+            set { SetBoolProperty(PropertyBool.IsArch, value); }
+        }
+
+        public bool? IsPsr
+        {
+            get { return GetBoolProperty(PropertyBool.IsPsr); }
+            set { SetBoolProperty(PropertyBool.IsPsr, value); }
+        }
+
+        public uint? Heritage
+        {
+            get { return GetIntProperty(PropertyInt.HeritageGroup); }
+            set { SetIntProperty(PropertyInt.HeritageGroup, value); }
+        }
+
+        public uint? Gender
+        {
+            get { return GetIntProperty(PropertyInt.Gender); }
+            set { SetIntProperty(PropertyInt.Gender, value); }
+        }
+
+        public uint? EyesTextureDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.EyesTexture); }
+            set { SetDataIdProperty(PropertyDataId.EyesTexture, value); }
+        }
+
+        public uint? DefaultEyesTextureDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.DefaultEyesTexture); }
+            set { SetDataIdProperty(PropertyDataId.DefaultEyesTexture, value); }
+        }
+
+        public uint? NoseTextureDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.NoseTexture); }
+            set { SetDataIdProperty(PropertyDataId.NoseTexture, value); }
+        }
+
+        public uint? DefaultNoseTextureDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.DefaultNoseTexture); }
+            set { SetDataIdProperty(PropertyDataId.DefaultNoseTexture, value); }
+        }
+
+        public uint? MouthTextureDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.MouthTexture); }
+            set { SetDataIdProperty(PropertyDataId.MouthTexture, value); }
+        }
+
+        public uint? DefaultMouthTextureDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.DefaultMouthTexture); }
+            set { SetDataIdProperty(PropertyDataId.DefaultMouthTexture, value); }
+        }
+
+        public uint? HairTextureDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.HairTexture); }
+            set { SetDataIdProperty(PropertyDataId.HairTexture, value); }
+        }
+
+        public uint? DefaultHairTextureDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.DefaultHairTexture); }
+            set { SetDataIdProperty(PropertyDataId.DefaultHairTexture, value); }
+        }
+
+        public uint? HeadObjectDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.HeadObject); }
+            set { SetDataIdProperty(PropertyDataId.HeadObject, value); }
+        }
+
+        public uint? SkinPaletteDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.SkinPalette); }
+            set { SetDataIdProperty(PropertyDataId.SkinPalette, value); }
+        }
+
+        public uint? HairPaletteDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.HairPalette); }
+            set { SetDataIdProperty(PropertyDataId.HairPalette, value); }
+        }
+
+        public uint? EyesPaletteDID
+        {
+            get { return GetDataIdProperty(PropertyDataId.EyesPalette); }
+            set { SetDataIdProperty(PropertyDataId.EyesPalette, value); }
+        }
+
+        public uint? Level
+        {
+            get { return GetIntProperty(PropertyInt.Level); }
+            set { SetIntProperty(PropertyInt.Level, value); }
+        }
+
+        public bool? GeneratorStatus
+        {
+            get { return GetBoolProperty(PropertyBool.GeneratorStatus); }
+            set { SetBoolProperty(PropertyBool.GeneratorStatus, value); }
+        }
+
+        public bool? GeneratorEnteredWorld
+        {
+            get { return GetBoolProperty(PropertyBool.GeneratorEnteredWorld); }
+            set { SetBoolProperty(PropertyBool.GeneratorEnteredWorld, value); }
+        }
+
+        public bool? GeneratorDisabled
+        {
+            get { return GetBoolProperty(PropertyBool.GeneratorDisabled); }
+            set { SetBoolProperty(PropertyBool.GeneratorDisabled, value); }
+        }
+
+        public bool? GeneratedTreasureItem
+        {
+            get { return GetBoolProperty(PropertyBool.GeneratedTreasureItem); }
+            set { SetBoolProperty(PropertyBool.GeneratedTreasureItem, value); }
+        }
+
+        public bool? GeneratorAutomaticDestruction
+        {
+            get { return GetBoolProperty(PropertyBool.GeneratorAutomaticDestruction); }
+            set { SetBoolProperty(PropertyBool.GeneratorAutomaticDestruction, value); }
+        }
+
+        public bool? CanGenerateRare
+        {
+            get { return GetBoolProperty(PropertyBool.CanGenerateRare); }
+            set { SetBoolProperty(PropertyBool.CanGenerateRare, value); }
+        }
+
+        public bool? CorpseGeneratedRare
+        {
+            get { return GetBoolProperty(PropertyBool.CorpseGeneratedRare); }
+            set { SetBoolProperty(PropertyBool.CorpseGeneratedRare, value); }
+        }
+
+        public bool? SuppressGenerateEffect
+        {
+            get { return GetBoolProperty(PropertyBool.SuppressGenerateEffect); }
+            set { SetBoolProperty(PropertyBool.SuppressGenerateEffect, value); }
+        }
+
+        public bool? ChestRegenOnClose
+        {
+            get { return GetBoolProperty(PropertyBool.ChestRegenOnClose); }
+            set { SetBoolProperty(PropertyBool.ChestRegenOnClose, value); }
+        }
+
+        public bool? ChestClearedWhenClosed
+        {
+            get { return GetBoolProperty(PropertyBool.ChestClearedWhenClosed); }
+            set { SetBoolProperty(PropertyBool.ChestClearedWhenClosed, value); }
+        }
+
+        public uint? GeneratorTimeType
+        {
+            get { return GetIntProperty(PropertyInt.GeneratorTimeType); }
+            set { SetIntProperty(PropertyInt.GeneratorTimeType, value); }
+        }
+
+        public uint? GeneratorProbability
+        {
+            get { return GetIntProperty(PropertyInt.GeneratorProbability); }
+            set { SetIntProperty(PropertyInt.GeneratorProbability, value); }
+        }
+
+        public uint? MaxGeneratedObjects
+        {
+            get { return GetIntProperty(PropertyInt.MaxGeneratedObjects); }
+            set { SetIntProperty(PropertyInt.MaxGeneratedObjects, value); }
+        }
+
+        public uint? GeneratorType
+        {
+            get { return GetIntProperty(PropertyInt.GeneratorType); }
+            set { SetIntProperty(PropertyInt.GeneratorType, value); }
+        }
+
+        public uint? ActivationCreateClass
+        {
+            get { return GetIntProperty(PropertyInt.ActivationCreateClass); }
+            set { SetIntProperty(PropertyInt.ActivationCreateClass, value); }
+        }
+
+        public bool? Ethereal
+        {
+            get { return GetBoolProperty(PropertyBool.Ethereal); }
+            set { SetBoolProperty(PropertyBool.Ethereal, value); }
+        }
+
+        public bool? IsOpen
+        {
+            get { return GetBoolProperty(PropertyBool.Open); }
+            set { SetBoolProperty(PropertyBool.Open, value); }
+        }
+
+        public bool? IsLocked
+        {
+            get { return GetBoolProperty(PropertyBool.Locked); }
+            set { SetBoolProperty(PropertyBool.Locked, value); }
+        }
+
+        public bool? DefaultLocked
+        {
+            get { return GetBoolProperty(PropertyBool.DefaultLocked); }
+            set { SetBoolProperty(PropertyBool.DefaultLocked, value); }
+        }
+
+        public bool? DefaultOpen
+        {
+            get { return GetBoolProperty(PropertyBool.DefaultOpen); }
+            set { SetBoolProperty(PropertyBool.DefaultOpen, value); }
+        }
+
+        public float? ResetInterval
+        {
+            get { return (float?)GetDoubleProperty(PropertyDouble.ResetInterval); }
+            set { SetDoubleProperty(PropertyDouble.ResetInterval, value); }
+        }
+
+        public double? ResetTimestamp
+        {
+            get { return GetDoubleProperty(PropertyDouble.ResetTimestamp); }
+            set { SetDoubleTimestamp(PropertyDouble.ResetTimestamp); }
+        }
+
+        public double? UseTimestamp
+        {
+            get { return GetDoubleProperty(PropertyDouble.UseTimestamp); }
+            set { SetDoubleTimestamp(PropertyDouble.UseTimestamp); }
+        }
+
+        public double? UseLockTimestamp
+        {
+            get { return GetDoubleProperty(PropertyDouble.UseLockTimestamp); }
+            set { SetDoubleTimestamp(PropertyDouble.UseLockTimestamp); }
+        }
+
+        public uint? LastUnlockerIID
+        {
+            get { return GetInstanceIdProperty(PropertyInstanceId.LastUnlocker); }
+            set { SetInstanceIdProperty(PropertyInstanceId.LastUnlocker, value); }
+        }
+
+        public string KeyCode
+        {
+            get { return GetStringProperty(PropertyString.KeyCode); }
+            set { SetStringProperty(PropertyString.KeyCode, value); }
+        }
+
+        public string LockCode
+        {
+            get { return GetStringProperty(PropertyString.LockCode); }
+            set { SetStringProperty(PropertyString.LockCode, value); }
+        }
+
+        public uint? ResistLockpick
+        {
+            get { return GetIntProperty(PropertyInt.ResistLockpick); }
+            set { SetIntProperty(PropertyInt.ResistLockpick, value); }
+        }
+
+        public uint? AppraisalLockpickSuccessPercent
+        {
+            get { return GetIntProperty(PropertyInt.AppraisalLockpickSuccessPercent); }
+            set { SetIntProperty(PropertyInt.AppraisalLockpickSuccessPercent, value); }
+        }
+
+        public Position Location
+        {
+            get { return GetPosition(PositionType.Location); }
+            set { SetPosition(PositionType.Location, value); }
+        }
+
+        public Position Destination
+        {
+            get { return GetPosition(PositionType.Destination); }
+            set { SetPosition(PositionType.Destination, value); }
+        }
+
+        public Position Instantiation
+        {
+            get { return GetPosition(PositionType.Instantiation); }
+            set { SetPosition(PositionType.Instantiation, value); }
+        }
+
+        public Position Sanctuary
+        {
+            get { return GetPosition(PositionType.Sanctuary); }
+            set { SetPosition(PositionType.Sanctuary, value); }
+        }
+
+        public Position Home
+        {
+            get { return GetPosition(PositionType.Home); }
+            set { SetPosition(PositionType.Home, value); }
+        }
+
+        public Position ActivationMove
+        {
+            get { return GetPosition(PositionType.ActivationMove); }
+            set { SetPosition(PositionType.ActivationMove, value); }
+        }
+
+        public Position Target
+        {
+            get { return GetPosition(PositionType.Target); }
+            set { SetPosition(PositionType.Target, value); }
+        }
+
+        public Position LinkedPortalOne
+        {
+            get { return GetPosition(PositionType.LinkedPortalOne); }
+            set { SetPosition(PositionType.LinkedPortalOne, value); }
+        }
+
+        public Position LastPortal
+        {
+            get { return GetPosition(PositionType.LastPortal); }
+            set { SetPosition(PositionType.LastPortal, value); }
+        }
+
+        public Position PortalStorm
+        {
+            get { return GetPosition(PositionType.PortalStorm); }
+            set { SetPosition(PositionType.PortalStorm, value); }
+        }
+
+        public Position CrashAndTurn
+        {
+            get { return GetPosition(PositionType.CrashAndTurn); }
+            set { SetPosition(PositionType.CrashAndTurn, value); }
+        }
+
+        public Position PortalSummonLoc
+        {
+            get { return GetPosition(PositionType.PortalSummonLoc); }
+            set { SetPosition(PositionType.PortalSummonLoc, value); }
+        }
+
+        public Position HouseBoot
+        {
+            get { return GetPosition(PositionType.HouseBoot); }
+            set { SetPosition(PositionType.HouseBoot, value); }
+        }
+
+        public Position LastOutsideDeath
+        {
+            get { return GetPosition(PositionType.LastOutsideDeath); }
+            set { SetPosition(PositionType.LastOutsideDeath, value); }
+        }
+
+        public Position LinkedLifestone
+        {
+            get { return GetPosition(PositionType.LinkedLifestone); }
+            set { SetPosition(PositionType.LinkedLifestone, value); }
+        }
+
+        public Position LinkedPortalTwo
+        {
+            get { return GetPosition(PositionType.LinkedPortalTwo); }
+            set { SetPosition(PositionType.LinkedPortalTwo, value); }
+        }
+
+        public Position Save1
+        {
+            get { return GetPosition(PositionType.Save1); }
+            set { SetPosition(PositionType.Save1, value); }
+        }
+
+        public Position Save2
+        {
+            get { return GetPosition(PositionType.Save2); }
+            set { SetPosition(PositionType.Save2, value); }
+        }
+
+        public Position Save3
+        {
+            get { return GetPosition(PositionType.Save3); }
+            set { SetPosition(PositionType.Save3, value); }
+        }
+
+        public Position Save4
+        {
+            get { return GetPosition(PositionType.Save4); }
+            set { SetPosition(PositionType.Save4, value); }
+        }
+
+        public Position Save5
+        {
+            get { return GetPosition(PositionType.Save5); }
+            set { SetPosition(PositionType.Save5, value); }
+        }
+
+        public Position Save6
+        {
+            get { return GetPosition(PositionType.Save6); }
+            set { SetPosition(PositionType.Save6, value); }
+        }
+
+        public Position Save7
+        {
+            get { return GetPosition(PositionType.Save7); }
+            set { SetPosition(PositionType.Save7, value); }
+        }
+
+        public Position Save8
+        {
+            get { return GetPosition(PositionType.Save8); }
+            set { SetPosition(PositionType.Save8, value); }
+        }
+
+        public Position Save9
+        {
+            get { return GetPosition(PositionType.Save9); }
+            set { SetPosition(PositionType.Save9, value); }
+        }
+
+        public Position RelativeDestination
+        {
+            get { return GetPosition(PositionType.RelativeDestination); }
+            set { SetPosition(PositionType.RelativeDestination, value); }
+        }
+
+        public Position TeleportedCharacter
+        {
+            get { return GetPosition(PositionType.TeleportedCharacter); }
+            set { SetPosition(PositionType.TeleportedCharacter, value); }
         }
 
         public uint? GetDataIdProperty(PropertyDataId property)